requirements.txt
git+https://github.com/Rapptz/discord.py
requests
discord.py
PyYAML
mysql-connector-python